+++ core/components/romanescobackyard/lexicon/nl/default.inc.php
@@ -14,6 +14,12 @@ $_lang['romanesco.menu.overview'] = 'Overzicht';
 $_lang['romanesco.menu.logo_title'] = 'Homepage van [[++site_name]]';
 $_lang['romanesco.menu.close_title'] = 'Sluiten';
 
+$_lang['romanesco.pagination.prev'] = 'Vorige pagina';
+$_lang['romanesco.pagination.next'] = 'Volgende pagina';
+$_lang['romanesco.pagination.first'] = 'Eerste pagina';
+$_lang['romanesco.pagination.last'] = 'Laatste pagina';
+$_lang['romanesco.pagination.load_more'] = 'Laad meer';
+
 // Social
 // ---------------------------------------------------------------------